Added homepage to the website.
Unlike other names, hugo does not allow for a content file to be called
index.md.  The proper means of creating a homepage without overriding
the default index in the theme in hugo is to specify an html file in
`layouts` that is called index.html.  Perhaps there is more elegant way
to generate this page, gut this is will work for now.